Hide the error text
This is a bitch. We *need* to set the error text to something. Otherwise the
widget doesn't think it's in an error state. See
https://github.com/flutter/flutter/blob/2d2a1ffec9/packages/flutter/lib/src/material/input_decorator.dart#L2072
But doing so introduces some height in which to accomadate the blank text. This
is undesirable. The hack here is to set the height of the TextStyle to zero. :S